requirements.txt
uvicorn
"fastapi[all]" 
"uvicorn[standard]"
PyJWT
python-jose
pandas 
PyPDF2
sentence-transformers
chromadb
groq
streamlit


uvicorn app.main:app --host 0.0.0.0 --port 8000 & 